Interaction with DB moved to separate service
